+++ Boa.Constrictor.Playwright/Tasks/Click.cs
@@ -1,21 +1,28 @@
-namespace Boa.Constrictor.Playwright.Tasks
+namespace Boa.Constrictor.Playwright
 {
     using System.Threading.Tasks;
-    using Boa.Constrictor.Playwright.Abilities;
-    using Boa.Constrictor.Playwright.Extensions;
     using Boa.Constrictor.Screenplay;
     using Microsoft.Playwright;
 
     /// <summary>
     /// Clicks a web element.
     /// </summary>
-    public class Click : AbstractPageTask
+    public class Click : AbstractLocatorTask
     {
-        private readonly string Selector;
+        private readonly LocatorClickOptions Options;
 
         #region Constructors
 
-        private Click(string selector) => this.Selector = selector;
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        /// <param name="locator">The target locator.</param>
+        /// <param name="options">Call options.</param>
+        private Click(IPlaywrightLocator locator, LocatorClickOptions options)
+            :base(locator)
+        {
+            Options = options;
+        }
 
         #endregion
 
@@ -24,9 +31,10 @@ namespace Boa.Constrictor.Playwright.Tasks
         /// <summary>
         /// Constructs the Task object.
         /// </summary>
-        /// <param name="selector">The target web elements Selector.</param>
+        /// <param name="locator">The target locator.</param>
+        /// <param name="options">Call options.</param>
         /// <returns></returns>
-        public static Click On(string selector) => new Click(selector);
+        public static Click On(IPlaywrightLocator locator, LocatorClickOptions options = null) => new Click(locator, options);
 
         #endregion
 
@@ -34,20 +42,19 @@ namespace Boa.Constrictor.Playwright.Tasks
 
         /// <summary>
         /// Clicks the web element.
-        /// Use browser actions instead of direct click (due to IE).
         /// </summary>
         /// <param name="actor">The Screenplay Actor.</param>
-        /// <param name="page">The current page.</param>
-        public override async Task PerformAsAsync(IActor actor, IPage page)
+        /// <param name="locator">The target locator.</param>
+        public override async Task PerformAsAsync(IActor actor, ILocator locator)
         {
-            await page.ClickAsync(this.Selector);
+            await locator.ClickAsync(Options);
         }
 
         /// <summary>
         /// Returns a description of the Task.
         /// </summary>
         /// <returns></returns>
-        public override string ToString() => $"click on {Selector}";
+        public override string ToString() => $"click on {Locator.Description}";
 
         #endregion
     }