+++ Boa.Constrictor.Playwright/Tasks/Go.cs
@@ -1,7 +1,6 @@
-namespace Boa.Constrictor.Playwright.Tasks
+namespace Boa.Constrictor.Playwright
 {
     using System.Threading.Tasks;
-    using Boa.Constrictor.Playwright.Extensions;
     using Boa.Constrictor.Screenplay;
     using Microsoft.Playwright;