+++ Boa.Constrictor.Playwright/CHANGELOG.md
@@ -10,7 +10,9 @@ Its format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 - Added project `Boa.Constrictor.Playwright` for items that improve [Playwright](https://playwright.dev/dotnet/) support
 - Added `BrowseTheWebSynchronously` ability to manage playwright, browsers, and pages
+- Added `PlaywrightLocator` to store description + selector logic of a playwright locator
+- Added `AbstractPageTask` that makes it easier to write Tasks that use the BrowseTheWebSynchronously Ability
+- Added `AbstractLocatorTask` to make it easier to write tasks that perform operations on a single locator
 - Added `Click`, `Fill`, and `Go` Tasks
 - Added `OpenNewPage` task to initialize page and navigate to the specified url
-- Added `AbstractPageTask` that makes it easier to write Tasks that use the BrowseTheWebSynchronously Ability
 - Added `Text` and `Attribute` questions